move format_rich to logging module
